pyproject.toml
[tool.poetry]
name = "fastapi-backend"
version = "0.1.0"
description = ""
authors = ["Manimegalan M <manimegalan@codewave.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.112.2"}
pydantic-settings = "^2.4.0"
python-decouple = "^3.8"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.33"}
loguru = "^0.7.2"
asyncpg = "^0.29.0"


[tool.poetry.group.dev.dependencies]
uvicorn = "^0.30.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
